Use int32 everywhere
On the amd64 architecture, this really matters. For some reason.

+++ loops/go/code.go
@@ -8,19 +8,19 @@ import (
 )
 
 func main() {
-	input, e := strconv.Atoi(os.Args[1]) // Get an input number from the command line
+	input, e := strconv.Atoi(os.Args[1])   // Get an input number from the command line
 	if e != nil {
 		panic(e)
 	}
-	u := input
-	r := int32(rand.Intn(10000))         // Get a random number 0 <= r < 10k
-	var a [10000]int32                   // Array of 10k elements initialized to 0
-	for i := 0; i < 10000; i++ {         // 10k outer loop iterations
-		k := a[i]                          // Use a local variable to loop in a register
-		for j := 0; j < 100000; j++ {      // 100k inner loop iterations, per outer loop iteration
-			k = k + int32(j%u)               // Simple sum
+	u := int32(input)
+	r := int32(rand.Intn(10000))           // Get a random number 0 <= r < 10k
+	var a [10000]int32                     // Array of 10k elements initialized to 0
+	for i := int32(0); i < 10000; i++ {    // 10k outer loop iterations
+		k := a[i]                            // Use a local variable to loop in a register
+		for j := int32(0); j < 100000; j++ { // 100k inner loop iterations, per outer loop iteration
+			k = k + j%u                        // Simple sum
 		}
-		a[i] += k + r                      // Add a random value to each element in array
+		a[i] += k + r                        // Add a random value to each element in array
 	}
-	fmt.Println(a[r])                    // Print out a single element from the array
+	fmt.Println(a[r])                      // Print out a single element from the array
 }
